refactor: 상대 시간 포맷 유틸 함수 분리
