+++ src/utils/formatDate.js
@@ -1,0 +1,18 @@
+export const formatDate = (dateString) => {
+  const createdDate = new Date(dateString);
+  const now = new Date();
+
+  const diffMs = now - createdDate;
+  const diffMinutes = Math.floor(diffMs / (1000 * 60));
+  const diffHours = Math.floor(diffMinutes / 60);
+
+  if (diffMinutes < 1) return '방금 전';
+  if (diffMinutes < 60) return `${diffMinutes}분 전`;
+  if (diffHours < 24) return `${diffHours}시간 전`;
+
+  return createdDate.toLocaleDateString('ko-KR', {
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+  });
+};
